Update README.md
XML Mapper v28.0.47991.2819

XML Mapper is the Swiss knife for transforming and mapping XML files.
This is an updated version of the XML Mapper that ships with RAD Studio.
At design time, it defines the mappings between generic XML documents
and the data packets that client datasets use. Each mapping describes
the correspondences between the nodes of an XML document and the
fields in a data packet. You can define mappings from an existing
XML schema (or document) to a client dataset that you define,
from an existing data packet to a new XML schema you define,
or between an existing XML schema and an existing data packet.

[Download XML Mapper v28.0.47991.2819](https://developer.team/delphi/35167-xml-mapper-v280479912819.html)